Improve accuracy of temperature conversion using floating-points.

+++ celsius.h
@@ -1,3 +1,3 @@
-int fromFahr(int fahrenheit);
+float fromFahr(float fahrenheit);
 void printFahrenheitRange(int upper, int lower, int skip);
 void promptConvertFahr();
